# src/main/java/com/example/userpages/model/CompanyPage.java
package com.example.userpages.model;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@Builder
@NoArgsConstructor
@AllArgsConstructor
public class CompanyPage {
    private String companyId;
    private String pageId;
}